requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pylint==2.6.0
pymongo==3.11.0
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
yarl==1.5.1
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
praw==7.1.0
prawcore==1.5.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
PyNaCl==1.4.0
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
toml==0.10.1
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.5.1aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
praw==7.1.0
prawcore==1.5.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
PyNaCl==1.4.0
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
spotipy==2.16.1
toml==0.10.1
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.5.1
